Updates for release 1.7.4 which pull in changes from upstream, and copies the Arduino Zero core changes into the zero_osc32k core

+++ cores/arduino_osc32k/itoa.h
@@ -26,7 +26,7 @@ extern "C"{
 
 extern char* itoa( int value, char *string, int radix ) ;
 extern char* ltoa( long value, char *string, int radix ) ;
-extern char* utoa( unsigned long value, char *string, int radix ) ;
+extern char* utoa( unsigned int value, char *string, int radix ) ;
 extern char* ultoa( unsigned long value, char *string, int radix ) ;
 
 #ifdef __cplusplus
